Add health check type support, load balancer enhancements, and end-to-end tests
- Introduce `healthCheckType` for services, supporting GRPC and HTTP health checks.
- Default health check type based on service protocol if unspecified.
- Implement configurable health check clients with TLS and protocol support.
- Enhance load balancer implementations (RoundRobinLB, LeastConnLB, WeightedRoundRobinLB) with circuit breaker state tracking and weighted target updates.
- Add comprehensive tests for health check handlers and gRPC health checks.
- Refactor `ProxyHandler` for clearer health check logic.
- Improve UI inputs for health check configuration and topology visualization.

+++ internal/telemetry/stats.go
@@ -79,7 +79,7 @@ func RecordPathRequest(host, path string, latencySeconds float64) {
 }
 
 // RecordTrace records a trace for an operation.
-func RecordTrace(id, operationName, serviceName string, durationMs int64, timestamp time.Time, status, path string) {
+func RecordTrace(id, operationName, serviceName string, durationMs float64, timestamp time.Time, status, path string) {
 	recordTraceToStore(id, operationName, serviceName, durationMs, timestamp, status, path)
 }
